decrease buffer_gb in audio interface test

+++ src/neuroconv/tools/hdmf.py
@@ -31,7 +31,7 @@ class GenericDataChunkIterator(HDMFGenericDataChunkIterator):
             # If the smallest full axis does not fit within the buffer size, form a square along the two smallest axes
             sub_square_buffer_shape = np.array(self.chunk_shape)
             if min(axis_sizes_bytes) > target_buffer_bytes:
-                k1 = np.floor((target_buffer_bytes / chunk_bytes) ** 0.5)
+                k1 = int(np.floor((target_buffer_bytes / chunk_bytes) ** 0.5))
                 for axis in [smallest_chunk_axis, second_smallest_chunk_axis]:
                     sub_square_buffer_shape[axis] = k1 * sub_square_buffer_shape[axis]
                 return tuple(sub_square_buffer_shape)
@@ -39,7 +39,7 @@ class GenericDataChunkIterator(HDMFGenericDataChunkIterator):
             smallest_chunk_axis = 0
             # Handle the case where the single axis is too large to fit in the buffer
             if axis_sizes_bytes[0] > target_buffer_bytes:
-                k1 = np.floor(target_buffer_bytes / chunk_bytes)
+                k1 = int(np.floor(target_buffer_bytes / chunk_bytes))
                 return tuple(
                     [
                         k1 * self.chunk_shape[0],
@@ -50,7 +50,7 @@ class GenericDataChunkIterator(HDMFGenericDataChunkIterator):
 
         # Original one-shot estimation has good performance for certain shapes
         chunk_to_buffer_ratio = buffer_gb * 1e9 / chunk_bytes
-        chunk_scaling_factor = np.floor(chunk_to_buffer_ratio ** (1 / num_axes))
+        chunk_scaling_factor = int(np.floor(chunk_to_buffer_ratio ** (1 / num_axes)))
         unpadded_buffer_shape = [
             np.clip(a=int(x), a_min=self.chunk_shape[j], a_max=self.maxshape[j])
             for j, x in enumerate(chunk_scaling_factor * np.array(self.chunk_shape))
@@ -74,7 +74,7 @@ class GenericDataChunkIterator(HDMFGenericDataChunkIterator):
                 buffer_bytes *= chunks_on_axis
                 padded_buffer_shape[axis] = self.maxshape[axis]
             else:  # Found an axis that is too large to use with the rest of the buffer; calculate how much can be used
-                k3 = np.floor(target_buffer_bytes / buffer_bytes)
+                k3 = int(np.floor(target_buffer_bytes / buffer_bytes))
                 padded_buffer_shape[axis] *= k3
                 break
         padded_buffer_bytes = np.prod(padded_buffer_shape) * self.dtype.itemsize
